refactor: split entities into jpa and domain
- move port package in application layer (integration package)
- adapt tests
- rewrite JPQL in native SQL query for discrepancy report
- change quantity type to Long and bigint in DDL

+++ src/main/resources/db/migration/V1__init_schema.sql
@@ -58,7 +58,7 @@ create table if not exists contents (
     id                  uuid         not null,
     container_unit_id   uuid         not null,
     sku                 varchar(255) not null,
-    quantity            integer      not null,
+    quantity            bigint      not null,
 
     constraint pk_contents primary key (id),
     constraint fk_contents_handling_unit
@@ -132,7 +132,7 @@ create table if not exists received_contents (
     id                uuid         not null,
     container_unit_id uuid,
     sku               varchar(255) not null,
-    quantity          integer      not null,
+    quantity          bigint      not null,
 
     constraint pk_received_contents primary key (id),
     constraint fk_received_contents_rhu
